Refactor IndexRelation classes.
