+++ engine/sql/core/src/main/scala/org/apache/spark/sql/index/IndexedRelation.scala
@@ -62,235 +62,3 @@ private[sql] abstract class IndexedRelation extends LogicalPlan {
 
   def withOutput(newOutput: Seq[Attribute]): IndexedRelation
 }
-
-private[sql] case class HashMapIndexedRelation(
-  output: Seq[Attribute],
-  child: SparkPlan,
-  table_name: Option[String],
-  column_keys: List[Attribute],
-  index_name: String)(var _indexedRDD: IndexRDD = null)
-  extends IndexedRelation with MultiInstanceRelation {
-  require(column_keys.length == 1)
-  val numShufflePartitions = child.sqlContext.conf.numShufflePartitions
-  val maxEntriesPerNode = child.sqlContext.conf.maxEntriesPerNode
-  val sampleRate = child.sqlContext.conf.sampleRate
-
-  if (_indexedRDD == null) {
-    buildIndex()
-  }
-
-  private[sql] def buildIndex(): Unit = {
-    val dataRDD = child.execute().map(row => {
-      val key = BindReferences.bindReference(column_keys.head, child.output).eval(row)
-      (key, row)
-    })
-
-    val partitionedRDD = HashPartition(dataRDD, numShufflePartitions)
-    val indexed = partitionedRDD.mapPartitions(iter => {
-      val data = iter.toArray
-      val index = HashMapIndex(data)
-      Array(IPartition(data.map(_._2), index)).iterator
-    }).persist(StorageLevel.MEMORY_AND_DISK_SER)
-
-    indexed.setName(table_name.map(n => s"$n $index_name").getOrElse(child.toString))
-    _indexedRDD = indexed
-  }
-
-  override def newInstance(): IndexedRelation = {
-    new HashMapIndexedRelation(output.map(_.newInstance()), child, table_name,
-      column_keys, index_name)(_indexedRDD).asInstanceOf[this.type]
-  }
-
-  override def withOutput(new_output: Seq[Attribute]): IndexedRelation = {
-    new HashMapIndexedRelation(new_output, child, table_name, column_keys, index_name)(_indexedRDD)
-  }
-
-  @transient override lazy val statistics = Statistics(
-    // TODO: Instead of returning a default value here, find a way to return a meaningful size
-    // estimate for RDDs. See PR 1238 for more discussions.
-    sizeInBytes = BigInt(child.sqlContext.conf.defaultSizeInBytes)
-  )
-}
-
-private[sql] case class TreeMapIndexedRelation(
-  output: Seq[Attribute],
-  child: SparkPlan,
-  table_name: Option[String],
-  column_keys: List[Attribute],
-  index_name: String)(var _indexedRDD: IndexRDD = null,
-                      var range_bounds: Array[Double] = null)
-  extends IndexedRelation with MultiInstanceRelation {
-  require(column_keys.length == 1)
-  require(column_keys.head.dataType.isInstanceOf[NumericType])
-  val numShufflePartitions = child.sqlContext.conf.numShufflePartitions
-  val maxEntriesPerNode = child.sqlContext.conf.maxEntriesPerNode
-  val sampleRate = child.sqlContext.conf.sampleRate
-
-  if (_indexedRDD == null) {
-    buildIndex()
-  }
-
-  private[sql] def buildIndex(): Unit = {
-    val dataRDD = child.execute().map(row => {
-      val key = BindReferences.bindReference(column_keys.head, child.output).eval(row)
-        .asInstanceOf[Number].doubleValue
-      (key, row)
-    })
-
-    val (partitionedRDD, tmp_bounds) = RangePartition.rowPartition(dataRDD, numShufflePartitions)
-    range_bounds = tmp_bounds
-    val indexed = partitionedRDD.mapPartitions(iter => {
-      val data = iter.toArray
-      val index = TreeMapIndex(data)
-      Array(IPartition(data.map(_._2), index)).iterator
-    }).persist(StorageLevel.MEMORY_AND_DISK_SER)
-
-    indexed.setName(table_name.map(n => s"$n $index_name").getOrElse(child.toString))
-    _indexedRDD = indexed
-  }
-
-  override def newInstance(): IndexedRelation = {
-    new TreeMapIndexedRelation(output.map(_.newInstance()), child, table_name,
-      column_keys, index_name)(_indexedRDD)
-      .asInstanceOf[this.type]
-  }
-
-  override def withOutput(new_output: Seq[Attribute]): IndexedRelation = {
-    new TreeMapIndexedRelation(new_output, child, table_name,
-      column_keys, index_name)(_indexedRDD, range_bounds)
-  }
-
-  @transient override lazy val statistics = Statistics(
-    // TODO: Instead of returning a default value here, find a way to return a meaningful size
-    // estimate for RDDs. See PR 1238 for more discussions.
-    sizeInBytes = BigInt(child.sqlContext.conf.defaultSizeInBytes)
-  )
-}
-
-private[sql] case class TreapIndexedRelation(
-  output: Seq[Attribute],
-  child: SparkPlan,
-  table_name: Option[String],
-  column_keys: List[Attribute],
-  index_name: String)(var _indexedRDD: IndexRDD = null,
-                     var range_bounds: Array[Double] = null)
-  extends IndexedRelation with MultiInstanceRelation {
-  require(column_keys.length == 1)
-  require(column_keys.head.dataType.isInstanceOf[NumericType])
-  val numShufflePartitions = child.sqlContext.conf.numShufflePartitions
-  val maxEntriesPerNode = child.sqlContext.conf.maxEntriesPerNode
-  val sampleRate = child.sqlContext.conf.sampleRate
-
-  if (_indexedRDD == null) {
-    buildIndex()
-  }
-
-  private[sql] def buildIndex(): Unit = {
-    val dataRDD = child.execute().map(row => {
-      val key = BindReferences.bindReference(column_keys.head, child.output).eval(row)
-        .asInstanceOf[Number].doubleValue
-      (key, row)
-    })
-
-    val (partitionedRDD, tmp_bounds) = RangePartition.rowPartition(dataRDD, numShufflePartitions)
-    range_bounds = tmp_bounds
-    val indexed = partitionedRDD.mapPartitions(iter => {
-      val data = iter.toArray
-      val index = Treap(data)
-      Array(IPartition(data.map(_._2), index)).iterator
-    }).persist(StorageLevel.MEMORY_AND_DISK_SER)
-
-    indexed.setName(table_name.map(n => s"$n $index_name").getOrElse(child.toString))
-    _indexedRDD = indexed
-  }
-
-  override def newInstance(): IndexedRelation = {
-    new TreapIndexedRelation(output.map(_.newInstance()), child, table_name,
-      column_keys, index_name)(_indexedRDD)
-      .asInstanceOf[this.type]
-  }
-
-  override def withOutput(new_output: Seq[Attribute]): IndexedRelation = {
-    new TreapIndexedRelation(new_output, child, table_name,
-      column_keys, index_name)(_indexedRDD, range_bounds)
-  }
-
-  @transient override lazy val statistics = Statistics(
-    // TODO: Instead of returning a default value here, find a way to return a meaningful size
-    // estimate for RDDs. See PR 1238 for more discussions.
-    sizeInBytes = BigInt(child.sqlContext.conf.defaultSizeInBytes)
-  )
-}
-
-private[sql] case class RTreeIndexedRelation(
-  output: Seq[Attribute],
-  child: SparkPlan,
-  table_name: Option[String],
-  column_keys: List[Attribute],
-  index_name: String)(var _indexedRDD: IndexRDD = null,
-                      var global_rtree: RTree = null)
-  extends IndexedRelation with MultiInstanceRelation {
-  private def checkKeys: Boolean = {
-    for (i <- column_keys.indices)
-      if (!(column_keys(i).dataType.isInstanceOf[DoubleType] ||
-        column_keys(i).dataType.isInstanceOf[IntegerType])) {
-        return false
-      }
-    true
-  }
-  require(checkKeys)
-
-  val numShufflePartitions = child.sqlContext.conf.numShufflePartitions
-  val maxEntriesPerNode = child.sqlContext.conf.maxEntriesPerNode
-  val sampleRate = child.sqlContext.conf.sampleRate
-  val transferThreshold = child.sqlContext.conf.transferThreshold
-
-  if (_indexedRDD == null) {
-    buildIndex()
-  }
-
-  private[sql] def buildIndex(): Unit = {
-    val dataRDD = child.execute().map(row => {
-      val now = column_keys.map(x =>
-        BindReferences.bindReference(x, child.output).eval(row).asInstanceOf[Number].doubleValue()
-      ).toArray
-      (new Point(now), row)
-    })
-
-    val dimension = column_keys.length
-    val max_entries_per_node = maxEntriesPerNode
-    val (partitionedRDD, mbr_bounds) = STRPartition(dataRDD, dimension, numShufflePartitions,
-      sampleRate, transferThreshold, max_entries_per_node)
-
-
-    val indexed = partitionedRDD.mapPartitions { iter =>
-      val data = iter.toArray
-      var index: RTree = null
-      if (data.length > 0) index = RTree(data.map(_._1).zipWithIndex, max_entries_per_node)
-      Array(IPartition(data.map(_._2), index)).iterator
-    }.persist(StorageLevel.MEMORY_AND_DISK_SER)
-
-    val partitionSize = indexed.mapPartitions(iter => iter.map(_.data.length)).collect()
-
-    global_rtree = RTree(mbr_bounds.zip(partitionSize)
-      .map(x => (x._1._1, x._1._2, x._2)), max_entries_per_node)
-    indexed.setName(table_name.map(n => s"$n $index_name").getOrElse(child.toString))
-    _indexedRDD = indexed
-  }
-
-  override def newInstance(): IndexedRelation = {
-    new RTreeIndexedRelation(output.map(_.newInstance()), child, table_name,
-      column_keys, index_name)(_indexedRDD).asInstanceOf[this.type]
-  }
-
-  override def withOutput(new_output: Seq[Attribute]): IndexedRelation = {
-    RTreeIndexedRelation(new_output, child, table_name,
-      column_keys, index_name)(_indexedRDD, global_rtree)
-  }
-
-  @transient override lazy val statistics = Statistics(
-    // TODO: Instead of returning a default value here, find a way to return a meaningful size
-    // estimate for RDDs. See PR 1238 for more discussions.
-    sizeInBytes = BigInt(child.sqlContext.conf.defaultSizeInBytes)
-  )
-}
